transfer

+++ verse_monster/train.py
@@ -131,7 +131,7 @@ def prep_dataset(dataset, keys_to_remove, num_datapoints_to_keep):
         return dataset
     dataset = limit_datset(dataset, num_datapoints_to_keep)
     remove_keys(dataset, keys_to_remove)
-    bos_tensor = torch.tensor([tokenizer.CharPhonemeTokenizer.bos_token_id])
+    bos_tensor = torch.tensor([tokenizer.CharPhonemeTokenizer.bos_token_id], dtype=torch.int)
     for dp in dataset:
         print('prep dp: ', dp)
         dp['labels'] = torch.cat([bos_tensor, dp['labels']])
